+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix emitting a comma when the only argument in a function is an ellipsis.
 - Fix function pointers inside function pointers, I think.
 - Fix sometimes being unable to demangle templated functions.
+- Fix emitting empty parenthesis on non-pointer arrays types.
 
 ## [0.1.0] - 2025-09-18